Mail.h is no longer automatically generated.
This tool uses Mail and the scripting bridge to send its emails, which
is weird but surprisingly painless. To get this to work, a Mail.h
header has to be generated from Mail's scripting dictionary. This can
be done on every build, but sadly, it didn't really work well - only on
the second build, and not at all for archiving. So I've added this
Mail.h to the repository instead and now just hope that no important
part of Mail's scripting architecture changes. At least not more than
Fanfiction.net