Refactored sheet support
Sheets are now handled by two window controllers: One for the main
window, one for the sheet (as base class of existing sheet
controllers). That means the NSApp beginSheet… stuff no longer has to
happen, error handling is nicer, and no more CFRetains.

There is a retain cycle here: The MainWindowController retains the
SheetController and vice versa. Either direction could easily be
broken, but practically speaking, it does not matter - the cycle goes
away when the sheet closes (the MWC removes itself as the parent of the
SC and sets its own SC to nil), and before that, there is a good reason
to keep both around.