Improved main menu; added resend
The main menu now makes a lot more sense for this kind of app. As an
important new feature, it is now possible to send stories again, no
matter when they were last updated. This will download all chapters
again. Since this is a somewhat rarely used feature (ideally), it's
only available via the menu. Also, some tiny changes so you can get the
main window again if it was closed.