Fixed localization; added accessibility descriptions
Apparently the project wasn't configured to properly generate the
string files that are used for localization. Sorry about that.

Also, the image buttons didn't have an accessibility description. This
was basically all that was standing between this app and full Voice
Over support, so I added them. But as it turns out,
DMLocalizedNibBundle.m does not support accessibility descriptions yet.