Loading images in overview now; auto-saving overview
Previously, the images for the stories never loaded correctly, because
setting valueURL to an URL of an image apparently didn't work, at least
not for a non-file URL. Now the StoryListEntry loads the image itself,
in the background.

As part of that, I also wanted the user's table layout preserved, so
this is implemented now. Took a checkbox, an autosave name and
identifiers for all columns.