Check cache for downloaded images before initiating new download.
The new download will, in practice, check the cache itself if necessary, but due to it working in the background and slowly on the main queue, there is a delay in the creation of the images. This easy way out avoids that. The cost? The code does not check whether the cache needs to be revalidated, so old images might stay around for a long time.