Changed chapter to same model as overview
Chapters now exist forever, not just as short wrappers, and update their contents. They are parts of the overview and have a (weak) reference to this overview. This is necessary for AO3, where there is more per-story state (in particular chapter URLs, which need to be stored explicitly), but also a cleaner design overall.